feat: add enablePrivateGit flag in IacTerragrunt struct
This commit introduces a flag `enablePrivateGit` in the IacTerragrunt struct
in `main.go`. It also updates several dependencies in `go.work.sum`, modifies
`dagger.json` to include a root source directory, and updates `go.mod` and
`go.sum` with new versions of `golang.org/x/exp` and `golang.org/x/sync`.

This new flag will allow users to enable or disable the private git
in the IacTerragrunt structure. It's still not fully implemented.