feat: Add basic functions for terragrunt intereoperability through Dagger (#50)
* feat: Add basic functions for terragrunt intereoperability through Dagger

Added functions to add environment variables to a container and convert a string
to a valid map of environment variables. Including error handling for invalid
environment variable formats.

* feat: Refactor WithSource method to include a workdir parameter

This change adds a workdir parameter to the WithSource method in the
Terragrunt struct. If no workdir is provided, it defaults to the
root workdir path. This allows for more flexibility in specifying
workdirs when mounting directories to the container.

* build: Remove unused functions in daggerx.go

Remove replaceEntryPointForShell, mergeSlices, addEnvVarsToContainer,
and toEnvVars functions from daggerx.go due to being unused. These
functions are not being utilized in the codebase.