feat: Refactor RunAll method in Terragrunt struct (#51)
Refactored the RunAll method in the Terragrunt struct to simplify the command
execution process. Introduced a new function `convertCMDToString` to convert
command slices to strings. Modified the method to use the full command as a string
and run it with the shell due to issues with the entrypoint. Added conversion of
environment variables to Dagger format for passing to the container.