fix: Disable automatic version bumping on push events
The changes made in this commit are to disable the automatic version bumping of the Dagger module on push events. This is done by commenting out the `push` event trigger in the workflow file. This change is made to prevent unnecessary version bumps during development and testing, and to only trigger the version bump when a pull request is closed.