feat: Add support for Dagger v0.12.5 in tests
The changes in this commit add the latest version of Dagger (v0.12.5) to the list of versions tested in the `TestDaggerWithDaggerCLI` function. This ensures that the test suite checks the compatibility of the module with the most recent version of Dagger, which is important for maintaining the reliability and robustness of the project.