feat: Update Dagger version to v0.12.5 in tests (#173)
* feat: Update Dagger version to v0.12.5 in tests

Updates the Dagger version used in the `TestDaggerSetupDaggerInDagger` test to the latest version, v0.12.5.

* feat: Update Dagger version to v0.12.5

Update the Dagger version used by the ModuleTemplate to v0.12.5. This version update is recommended to take advantage of the latest features and bug fixes in the Dagger engine.

* feat: Add support for Dagger v0.12.5 in tests

The changes in this commit add the latest version of Dagger (v0.12.5) to the list of versions tested in the `TestDaggerWithDaggerCLI` function. This ensures that the test suite checks the compatibility of the module with the most recent version of Dagger, which is important for maintaining the reliability and robustness of the project.

* feat: Update dagger.go tests

  
This commit updates the dagger.go tests in the module-template project. The main changes are:

- Removed the check for the "registry.dagger.io/engine" string in the docker ps output, as this check may not be necessary and could be brittle.
- Improved the formatting of the "dagger version" command execution to make the code more readable.

The goal of these changes is to make the tests more robust and maintainable.

* feat: Update Dagger Engine version to v0.12.5

- Update the Dagger Engine version used in the module and tests from v0.12.4 to v0.12.5
- Remove the check for the presence of "registry.dagger.io/engine" in the docker ps output, as it may not be necessary for the test
- Update the list of Dagger versions tested in the `TestDaggerWithDaggerCLI` test to include v0.12.5