feat: Add base image configuration for ModuleTemplate
This commit introduces three new methods to the `ModuleTemplate` struct:

1. `BaseAlpine`: Sets the base image to an Alpine Linux image and creates the base container. The version of the Alpine image can be specified, and it defaults to "latest" if not provided.

2. `BaseUbuntu`: Sets the base image to an Ubuntu Linux image and creates the base container. The version of the Ubuntu image can be specified, and it defaults to "latest" if not provided.

3. `BaseBusyBox`: Sets the base image to a BusyBox Linux image and creates the base container. The version of the BusyBox image can be specified, and it defaults to "latest" if not provided.

These methods provide a convenient way to set the base image for the `ModuleTemplate` and ensure that the container is created with the desired base image. This change adds flexibility and allows users to choose the base image that best suits their needs.