feat: Use a default image version for base containers
The changes in this commit update the default image version used for the base containers in the `ModuleTemplate` struct. Previously, the `version` parameter was set to `"latest"` if an empty string was provided. This commit introduces a new constant `defaultImageVersionLatest` and updates the `BaseAlpine`, `BaseUbuntu`, and `BaseBusyBox` methods to use this constant instead of hardcoding `"latest"`.

This change ensures consistency in the default image version used across the different base container methods, making the code more maintainable and easier to update in the future.