feat: Add test for overriding container in Go toolbox
The changes add a new test function `TestgotoolboxWithOverridingContainer` that verifies the ability to override the container used in the Go toolbox. The test performs the following steps:

1. Initializes the Go toolbox with a custom container based on the `ubuntu:22.04` image.
2. Installs the Go programming language on the custom container using `apt-get`.
3. Checks that the Go version is installed correctly by verifying the output of the `go version` command.

This new test ensures that the Go toolbox can be used with a custom container, allowing for greater flexibility in the development environment.